Split middleware routing helpers
